set_initial_no_auth_action_count should also set request_count to 0.
